requirements.txt
# Runtime
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
llama-cpp-python>=0.3.0
huggingface-hub>=0.27.0
aiosqlite>=0.22.0

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
ruff>=0.8.0
mypy>=1.13.0
